feat(#44): Add support for CST and CSV drive control

+++ libs/AutomationWare/inc/aw_j_series.h
@@ -39,7 +39,7 @@ int32_t ec_get_actpos_wrd_aw_j_series(uint16_t drive);
 
 int32_t ec_get_actvel_wrd_aw_j_series(uint16_t drive);
 
-int16_t ec_get_acttorq_wrd_aw_j_series(uint16_t drive);
+int32_t ec_get_acttorq_wrd_aw_j_series(uint16_t drive);
 
 gberror_t ec_set_setpos_wrd_aw_j_series(uint16_t drive, int32_t setpos);
